server/requirements.txt
aiohttp
numpy
mysql-connector-python